rename file name of gridwise batched gemm
